requirements.txt
# Core dependencies
torch>=2.0.0
transformers>=4.35.0
accelerate>=0.25.0
bitsandbytes>=0.41.0  # For 8-bit and 4-bit quantization
numpy>=1.24.0
matplotlib>=3.7.0
Pillow>=9.4.0
tqdm>=4.65.0

# SVG processing
picosvg>=0.20.0
cairosvg>=2.7.0  # For SVG rendering (optional)
svglib>=1.5.1    # Alternative SVG renderer

# Web API
flask>=2.3.0
gunicorn>=21.2.0  # For production deployment
waitress>=2.1.2   # Alternative WSGI server

# Data handling and storage
pandas>=2.0.0
h5py>=3.9.0

# Utilities
scipy>=1.11.0
scikit-learn>=1.3.0
jsonlines>=3.1.0
python-dotenv>=1.0.0